feat: added git diff & moved common functions to utils
