+++ src/commands/dev/config/push.rs
@@ -1,21 +1,17 @@
 use crate::error::DevConfigError;
 use crate::loader::new_spinner_progress_bar;
 use crate::utils::annotations::VaultSecretAnnotation;
-use crate::{
-    error::CliError, services::vault::Vault, utils::annotations::read_vault_annotation,
-    utils::line::Line,
-};
-use dialoguer::console::{style, Style};
+use crate::{error::CliError, services::vault::Vault};
 use dialoguer::Confirm;
 use dialoguer::{theme::ColorfulTheme, Select};
-use ignore::{overrides::OverrideBuilder, WalkBuilder};
-use log::debug;
 use owo_colors::OwoColorize;
-use similar::{ChangeTag, TextDiff};
 use std::collections::HashMap;
-use std::{
-    env::current_dir,
-    path::{Path, PathBuf},
+use std::path::PathBuf;
+
+use super::diff::print_diff;
+use super::utils::{
+    filter_config_with_secret_annotations, get_available_files, get_updated_configs,
+    remove_parent_directories,
 };
 
 pub async fn handle_config_push() -> Result<bool, CliError> {
@@ -56,7 +52,7 @@ pub async fn handle_config_push() -> Result<bool, CliError> {
             "{}",
             "There is only one config file to update...".bright_yellow()
         );
-        let (config_path, annotation) = updated_configs.iter().next().unwrap();
+        let (config_path, (annotation, _, _)) = updated_configs.iter().next().unwrap();
 
         update_secrets(
             &vault,
@@ -70,50 +66,6 @@ pub async fn handle_config_push() -> Result<bool, CliError> {
     }
 
     Ok(true)
-}
-
-async fn get_updated_configs(
-    vault: &Vault,
-    vault_token: &str,
-    config_files: &HashMap<String, VaultSecretAnnotation>,
-) -> Result<HashMap<String, VaultSecretAnnotation>, CliError> {
-    // Comparing local vs remote ....
-    println!("{}", "comparing local config vs remote config...".cyan());
-
-    let mut updated_configs = HashMap::new();
-
-    for config_file in config_files {
-        let (config_path, vault_secret_annotation) = config_file;
-        let remote_secrets = vault
-            .get_secrets(vault_token, &vault_secret_annotation.secret_path)
-            .await?
-            .data;
-
-        let config_string =
-            get_local_config_as_string(&vault_secret_annotation.destination_file, config_path)
-                .map_err(|error| {
-                    debug!("Error: {:?}", error);
-                    CliError::DevConfigError(DevConfigError::ConfigSecretNotFound)
-                })?;
-
-        // Get only one key from hashmap
-        let remote_config = match remote_secrets.get(&vault_secret_annotation.secret_name) {
-            Some(config) => config,
-            None => {
-                return Err(CliError::DevConfigError(
-                    DevConfigError::InvalidSecretPath {
-                        config_path: remove_parent_directories(config_path),
-                    },
-                ));
-            }
-        };
-
-        if has_diff(remote_config, &config_string) {
-            updated_configs.insert(config_path.clone(), vault_secret_annotation.clone());
-        }
-    }
-
-    Ok(updated_configs)
 }
 
 fn get_local_config_as_string(
@@ -176,22 +128,14 @@ async fn update_secrets(
     Ok(())
 }
 
-fn has_diff(old_secret_config: &str, new_secret_config: &str) -> bool {
-    let changeset = TextDiff::from_lines(old_secret_config, new_secret_config);
-
-    changeset
-        .iter_all_changes()
-        .any(|change| matches!(change.tag(), ChangeTag::Delete | ChangeTag::Insert))
-}
-
 async fn select_config(
-    available_config: &HashMap<String, VaultSecretAnnotation>,
+    available_config: &HashMap<String, (VaultSecretAnnotation, String, String)>,
 ) -> (String, VaultSecretAnnotation) {
     let selection = Select::with_theme(&ColorfulTheme::default())
         .items(
             &available_config
                 .iter()
-                .map(|(config_path, annotation)| {
+                .map(|(config_path, (annotation, _, _))| {
                     format!(
                         "{} \t {}::{}/{}#{}",
                         remove_parent_directories(config_path),
@@ -214,7 +158,7 @@ async fn select_config(
 
     return match selection {
         Some(index) => {
-            let (config_path, annotation) = available_config.iter().nth(index).unwrap();
+            let (config_path, (annotation, _, _)) = available_config.iter().nth(index).unwrap();
             (config_path.clone(), annotation.clone())
         }
         None => {
@@ -222,127 +166,4 @@ async fn select_config(
             std::process::exit(0);
         }
     };
-}
-
-fn get_available_files() -> Result<Vec<PathBuf>, CliError> {
-    let current_path = current_dir()?;
-    let available_files = get_dev_config_files(&current_path);
-
-    Ok(available_files)
-}
-
-fn filter_config_with_secret_annotations(
-    available_files: Vec<PathBuf>,
-) -> Result<HashMap<String, VaultSecretAnnotation>, CliError> {
-    let mut filtered_annotations: HashMap<String, VaultSecretAnnotation> = HashMap::new();
-
-    for file in available_files {
-        let file_contents = std::fs::read_to_string(file.clone())?;
-        let annotations = read_vault_annotation(&file_contents);
-
-        for annotation in annotations {
-            if annotation.key == "wukong.mindvalley.dev/config-secrets-location"
-                && annotation.source == "vault"
-                && annotation.engine == "secret"
-            {
-                filtered_annotations.insert(file.to_string_lossy().to_string(), annotation);
-            }
-        }
-    }
-
-    Ok(filtered_annotations)
-}
-
-fn print_diff(old_secret_config: &str, new_secret_config: &str, secret_path: &str) {
-    println!();
-    println!("{}", secret_path.dimmed());
-
-    let diff = TextDiff::from_lines(old_secret_config, new_secret_config);
-
-    for (idx, group) in diff.grouped_ops(3).iter().enumerate() {
-        if idx > 0 {
-            println!("{:-^1$}", "-", 80);
-        }
-        for op in group {
-            for change in diff.iter_inline_changes(op) {
-                let (sign, s) = match change.tag() {
-                    ChangeTag::Delete => ("-", Style::new().red()),
-                    ChangeTag::Insert => ("+", Style::new().green()),
-                    ChangeTag::Equal => (" ", Style::new().dim()),
-                };
-                print!(
-                    "{}{} |{}",
-                    style(Line(change.old_index())).dim(),
-                    style(Line(change.new_index())).dim(),
-                    s.apply_to(sign).bold(),
-                );
-                for (emphasized, value) in change.iter_strings_lossy() {
-                    if emphasized {
-                        print!("{}", s.apply_to(value).underlined().on_black());
-                    } else {
-                        print!("{}", s.apply_to(value));
-                    }
-                }
-                if change.missing_newline() {
-                    println!();
-                }
-            }
-        }
-    }
-
-    println!();
-}
-
-fn remove_parent_directories(path: &str) -> String {
-    let current_dir = current_dir().unwrap();
-    let path = Path::new(path);
-
-    path.strip_prefix(current_dir)
-        .map(|p| p.to_owned())
-        .unwrap_or(path.to_owned())
-        .into_os_string()
-        .into_string()
-        .unwrap()
-}
-
-fn get_dev_config_files(path: &Path) -> Vec<PathBuf> {
-    let mut overrides = OverrideBuilder::new(path);
-    overrides.add("**/config/dev.exs").unwrap();
-
-    WalkBuilder::new(path)
-        .overrides(overrides.build().unwrap())
-        .build()
-        .flatten()
-        .filter(|e| e.path().is_file())
-        .map(|e| e.path().to_path_buf())
-        .collect()
-}
-
-// Test:
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_remove_parent_directories() {
-        let current_dir = std::env::current_dir().unwrap();
-
-        let path = current_dir.join("some_directory/some_file.txt");
-        let input = path.to_str().unwrap();
-        let expected_output = "some_directory/some_file.txt";
-        assert_eq!(remove_parent_directories(input), expected_output);
-
-        let path = current_dir.join("another_directory/another_file.txt");
-        let input = path.to_str().unwrap();
-        let expected_output = "another_directory/another_file.txt";
-        assert_eq!(remove_parent_directories(input), expected_output);
-
-        let input = "/absolute/path/to/some/file.txt";
-        let expected_output = "/absolute/path/to/some/file.txt";
-        assert_eq!(remove_parent_directories(input), expected_output);
-
-        let input = "relative/path/to/some/file.txt";
-        let expected_output = "relative/path/to/some/file.txt";
-        assert_eq!(remove_parent_directories(input), expected_output);
-    }
 }
